Added modules util.js and ad.js
