+++ js/util.js
@@ -1,0 +1,25 @@
+const getRundomIntNumber = (first, second) => {
+  const lower = Math.ceil(Math.min(Math.abs(first), Math.abs(second)));
+  const upper = Math.floor(Math.max(Math.abs(first), Math.abs(second)));
+  return Math.floor(Math.random() * (upper - lower + 1) + lower);
+};
+
+const getRandomFloatNumber = (first, second, pointPlace) => {
+  const lower = Math.ceil(Math.min(Math.abs(first), Math.abs(second)));
+  const upper = Math.floor(Math.max(Math.abs(first), Math.abs(second)));
+  return Number((Math.random() * (upper - lower) + lower).toFixed(pointPlace));
+};
+
+const getRandomArrayElement = (array, minValue) => {
+  const elementIndex = getRundomIntNumber(minValue, array.length - 1);
+  return array[elementIndex];
+};
+
+const getRandomArrayLength = (array, minValue) => {
+  const newArray = array.slice();
+  const newArrayLength = getRundomIntNumber(minValue, newArray.length);
+  newArray.length = newArrayLength;
+  return newArray;
+};
+
+export { getRundomIntNumber, getRandomFloatNumber, getRandomArrayElement, getRandomArrayLength };
